Extract types.
